fix: ui fixes
