+++ src/app/login/components/Login.test.tsx
@@ -1,7 +1,7 @@
 import { expect, test, vi } from "vitest";
 import { fireEvent, render, screen, waitFor } from "@testing-library/react";
 import { LoginPage } from "./Login";
-import { loginUser } from "@/app/members/actions/action";
+import { loginUser } from "@/app/members/actions/loginUser";
 
 const credentialsMock = {
   phoneNumber: "923000000000",
