+++ src/app/members/components/MemberFormBuilder.test.tsx
@@ -1,7 +1,8 @@
 import { render, screen, fireEvent, waitFor } from "@testing-library/react";
 import { describe, expect, it, vi } from "vitest";
 import { MemberFormBuilder } from "./MemberFormBuilder";
-import { createUser } from "@/app/members/actions/action";
+import { createUser } from "../actions/createUser";
+
 
 
 type TestElement = Document | Element | Window | Node;
